requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
tqdm>=4.64.0
pyyaml>=6.0
google-genai>=1.5.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0 